+++ src/api/endpoints/execute/get-execute-list/v3.ts
@@ -102,12 +102,12 @@ export const getExecuteListV3Options: RouteOptions = {
           listingTime: Joi.string()
             .pattern(regex.unix_timestamp)
             .description(
-              "Unix timestamp indicating when listing will be listed. Example: `1656080318`"
+              "Unix timestamp (seconds) indicating when listing will be listed. Example: `1656080318`"
             ),
           expirationTime: Joi.string()
             .pattern(regex.unix_timestamp)
             .description(
-              "Unix timestamp indicating when listing will expire. Example: `1656080318`"
+              "Unix timestamp (seconds) indicating when listing will expire. Example: `1656080318`"
             ),
           salt: Joi.string()
             .pattern(regex.number)